bug fixes on #67
